requirements.txt

streamlit
openai>=1.35.0
python-dotenv
langchain

PyMuPDF
pillow
pandas

azure-search-documents>=11.4.0
azure-core>=1.29.1
azure-cognitiveservices-vision-computervision
msrest
